Remove redundant imports and add helper function
